remove add to-many assoc in DDM, including cenz adapter

+++ SIAGRO/instance_2_ne014/graphql-server/resolvers/location.js
@@ -91,6 +91,54 @@ location.prototype.accessionsConnection = function({
     };
 }
 
+
+/**
+ * countAllAssociatedRecords - Count records associated with another given record
+ *
+ * @param  {ID} id      Id of the record which the associations will be counted
+ * @param  {objec} context Default context by resolver
+ * @return {Int}         Number of associated records
+ */
+async function countAllAssociatedRecords(id, context) {
+
+    let location = await resolvers.readOneLocation({
+        locationId: id
+    }, context);
+    //check that record actually exists
+    if (location === null) throw new Error(`Record with ID = ${id} does not exist`);
+    let promises_to_many = [];
+    let promises_to_one = [];
+
+    promises_to_many.push(location.countFilteredAccessions({}, context));
+
+    let result_to_many = await Promise.all(promises_to_many);
+    let result_to_one = await Promise.all(promises_to_one);
+
+    let get_to_many_associated = result_to_many.reduce((accumulator, current_val) => accumulator + current_val, 0);
+    let get_to_one_associated = result_to_one.filter((r, index) => r !== null).length;
+
+    return get_to_one_associated + get_to_many_associated;
+}
+
+/**
+ * validForDeletion - Checks wether a record is allowed to be deleted
+ *
+ * @param  {ID} id      Id of record to check if it can be deleted
+ * @param  {object} context Default context by resolver
+ * @return {boolean}         True if it is allowed to be deleted and false otherwise
+ */
+async function validForDeletion(id, context) {
+    if (await countAllAssociatedRecords(id, context) > 0) {
+        throw new Error(`Accession with accession_id ${id} has associated records and is NOT valid for deletion. Please clean up before you delete.`);
+    }
+
+    if (context.benignErrors.length > 0) {
+        throw new Error('Errors occurred when counting associated records. No deletion permitted for reasons of security.');
+    }
+
+    return true;
+}
+
 module.exports = {
 
     /**
@@ -108,11 +156,6 @@ module.exports = {
         order,
         pagination
     }, context) {
-        /**
-         * Debug
-         */
-        console.log("\n-@--resolver: on: locationsConnection");
-
         //check: adapters
         let registeredAdapters = Object.values(location.registeredAdapters);
         if (registeredAdapters.length === 0) {
@@ -165,11 +208,6 @@ module.exports = {
     readOneLocation: async function({
         locationId
     }, context) {
-        /**
-         * Debug
-         */
-        console.log("\n-@--resolver: on: readOneLocation");
-
         //check: adapters auth
         try {
             let authorizationCheck = await checkAuthorization(context, location.adapterForIri(locationId), 'read');
@@ -192,11 +230,6 @@ module.exports = {
      * @return {object}         New record created
      */
     addLocation: async function(input, context) {
-        /**
-         * Debug
-         */
-        console.log("\n-@--resolver: on: addLocation");
-
         //check: input has idAttribute
         if (!input.locationId) {
             throw new Error(`Illegal argument. Provided input requires attribute 'locationId'.`);
@@ -246,16 +279,13 @@ module.exports = {
     deleteLocation: async function({
         locationId
     }, context) {
-        /**
-         * Debug
-         */
-        console.log("\n-@--resolver: on: deleteLocation");
-
         //check: adapters auth
         try {
             let authorizationCheck = await checkAuthorization(context, location.adapterForIri(locationId), 'delete');
             if (authorizationCheck === true) {
-                return location.deleteOne(locationId);
+                if (await location.validForDeletion(locationId, context)) {
+                    return location.deleteOne(locationId);
+                }
             } else { //adapter not auth
                 throw new Error("You don't have authorization to perform this action on adapter");
             }
@@ -273,11 +303,6 @@ module.exports = {
      * @return {object}         Updated record
      */
     updateLocation: async function(input, context) {
-        /**
-         * Debug
-         */
-        console.log("\n-@--resolver: on: updateLocation");
-
         //check: input has idAttribute
         if (!input.locationId) {
             throw new Error(`Illegal argument. Provided input requires attribute 'locationId'.`);
@@ -308,11 +333,6 @@ module.exports = {
     countLocations: async function({
         search
     }, context) {
-        /**
-         * Debug
-         */
-        console.log("\n-@--resolver: on: countLocation");
-
         //check: adapters
         let registeredAdapters = Object.values(location.registeredAdapters);
         if (registeredAdapters.length === 0) {
